requirements.txt
pillow==10.0.0
requests==2.31.0
pyTelegramBotAPI==4.12.0
python-dotenv==1.0.0
